requirements.txt

flatbuffers
numpy
h5py
confluent-kafka
kafka-python
pyzmq
xmltodict
ess-streaming-data-types
